+++ libs/components/src/index.ts
@@ -2,6 +2,9 @@ export * from './constants/navigation';
 export * from './lib/AccountRemoveSection';
 export * from './lib/admin/AdminAccountList';
 export * from './lib/admin/AdminBusinessRegistrationList';
+export * from './lib/admin/AdminGoodsConfirmationDialog';
+export * from './lib/admin/AdminGoodsList';
+export * from './lib/admin/AdminGoodsStatusButtons';
 export * from './lib/admin/AdminLoginForm';
 export * from './lib/admin/AdminPageLayout';
 export * from './lib/BusinessRegistrationBox';
@@ -33,6 +36,7 @@ export * from './lib/GoodsRegistForm';
 export * from './lib/GoodsStatusBadge';
 export * from './lib/HorizontalImageGallery';
 export * from './lib/LoginForm';
+export * from './lib/MainBeta';
 export * from './lib/MypageLayout';
 export * from './lib/Navbar';
 export * from './lib/OrderDetailActions';
@@ -68,11 +72,4 @@ export * from './lib/SocialAccountUnlinkSection';
 export * from './lib/SummaryList';
 export * from './lib/TestComponent';
 export * from './lib/TextDotConnector';
-export * from './lib/admin/AdminPageLayout';
-export * from './lib/admin/AdminLoginForm';
-export * from './lib/admin/AdminAccountList';
-export * from './lib/admin/AdminBusinessRegistrationList';
-export * from './lib/admin/AdminGoodsList';
-export * from './lib/admin/AdminGoodsConfirmationDialog';
 export * from './lib/TextViewerWithDetailModal';
-export * from './lib/admin/AdminGoodsStatusButtons';
